Always output the resumptionToken tag.
This adds again the resumptionToken if we have a complete list without a
token itself. This is not a must feature but quite common and I think
it's not forbidden.

This changes the XSLT stylesheet a bit because we output the "cursor" in
the bottom which is not the amount of submitted records but the position
in the list of records.

Just another extample of a complete list:

https://www.tib.eu/oai/public/repository/open?verb=ListIdentifiers&metadataPrefix=oai_dc&set=zbmql